Added editing elments of list

+++ src/containers/Layout/ToDoList/List/index.js
@@ -1,28 +1,51 @@
 import React, { Component } from 'react';
 
+import EditValue from './componets/EditValue/index'
+
 class List extends Component{
 
     render() {
-        const {data, deleteItem} = this.props;
+        const {data, deleteItem, editItem, submitEdit} = this.props;
 
         return(
                 <div>
                     <ul>
                         {
-                            data.map((item, index) => (
-                                <li key={index}>{item}
-                                <span
-                                onClick={()=>deleteItem(index)}>
-                                    DELETE
-                                </span>
-                                    </li>
+                            data.map((item, index) => {
+                                if (item.needEdit) {
+                                    return(
+                                        <li
+                                        key={index}>
+                                            <EditValue
+                                                value={item.value}
+                                                submit={submitEdit}
+                                                index={index}
+                                            />
+                                        </li>
+
+                                    )
+                                }
+                                else
+                                {
+                                    return( <li key={index}>{item.value}
+                                        <button
+                                            onClick={() => deleteItem(index)}>
+                                            Delete
+                                        </button>
+                                        <button
+                                            onClick={() => editItem(index)}>
+                                            Edit
+                                        </button>
+                                    </li>)
+                                }
 
 
-                            ))
+
+                            })
                         }
 
                     </ul>
-                    <img src="../../../../icons/deleteElement.jpg" width="50px" height="50px" alt=""/>
+
                 </div>
 
         )
